# bootstrap: Link LLVM as a dylib with ThinLTO