--- a/src/librustc_llvm/build.rs
+++ b/src/librustc_llvm/build.rs
@@ -142,6 +142,10 @@ fn main() {
             continue;
         }
 
+        if flag.starts_with("-flto") {
+            continue;
+        }
+
         // -Wdate-time is not supported by the netbsd cross compiler
         if is_crossed && target.contains("netbsd") && flag.contains("date-time") {
             continue;