… src/test/run-make-fulldeps/llvm-pass/Makefile
… src/test/run-make-fulldeps/llvm-pass/llvm-function-pass.so.cc
… src/test/run-make-fulldeps/llvm-pass/llvm-module-pass.so.cc
… src/test/run-make-fulldeps/llvm-pass/main.rs
… src/test/run-make-fulldeps/llvm-pass/plugin.rs