@alexcrichton 2018-12-19
Ok, so here's some information about this PR now:

The first attempt, found in just the first commit, changes the Linux dist builders to build LLVM with ThinLTO but then links rustc against the dynamic library produced. This means that rustc doesn't rerun ThinLTO throughout the stages. It also means, though, that we need to copy around the LLVM dynamic library when creating the sysroot. This attempt's [benchmarks were pretty bad](https://perf.rust-lang.org/compare.html?start=041254b81495a5aa67af839e00b890e78ed0cbeb&end=a454167f2914c0ef3e34f7b63654485f8e276b78), though. Local profiling showed that this was most likely due to the fact that we [use `RTLD_NOW` when opening the codegen backend](https://github.com/rust-lang/rust/blob/d99a320cba42f661aebfa1293b7b2ec3603dda75/src/librustc_driver/lib.rs#L222-L228) and the dynamic linker was processing a whole lot more symbols (all in LLVM).

The second attempt, outlined in the second commit, eschews `RTLD_NOW` and instead just uses the normal `open` that we've implemented. Benchmarks [then confirmed that this change doens't regress performance](https://perf.rust-lang.org/compare.html?start=cb84844e83ee88684ef89cc02221a26abbf92530&end=a8b3bd70936e9ce4e4db4aaf7b4a4cae6b8f3d39), as expected. Apparently the dynamic linker can be a sizable portion of our runtimes! Probably not a good sign that we have so many dynamic libraries... In any case this commit broke tests because `RTLD_NOW` was explicitly added to enusre that our one test for an LLVM plugin kept working.

The current PR, outlined in the final commit, deletes the test that we can load an LLVM plugin at runtime. I feel like the writing has been on the wall for this test for quite some time now. LLVM plugins were always unstable and AFAIK aren't even used at all any more. When they *could* be used, they were extremely difficult to use.

Technically this PR is a regression in unstable functionality, but I personally feel that it's worth it. Taking a look at the [try build](https://travis-ci.org/rust-lang/rust/builds/469715608?utm_source=github_status&utm_medium=notification) compared against the [previous auto build](https://travis-ci.org/rust-lang/rust/builds/469249414) we see:

|  builder | time before | time after |
|---------|------------|-------------|
| dist-x86_64-linux | 2h13m | 1h45m |
| dist-x86_64-linux-alt | 2h19m | 2h1m |

Here we're seeing 20+ minute reductions in build time for these builders, thanks to the fewer runs of ThinLTO we're having them do. That alone is worth it I believe because these builders are some of the slowest on Travis.