@Mark-Simulacrum 2018-12-19
I agree that we should go ahead and do this; LLVM plugins are not something that we need to support and the wins are substantial. Plus, if necessary, we can revert.

I would also be interested in seeing if we could investigate the dynamic linker performance further to see if we might be able to get away with loading less symbols overall and perhaps speed up tests...