@michaelwoerister 2018-12-19
I'm also in favor of moving forward with this. The slight regression we see now is small compared to what ThinLTO'd LLVM gained us in the first place and this change should allow us to enable ThinLTO for LLVM on the other major platforms (right?). 

cc @rust-lang/wg-compiler-performance (so the slowdown for small crates doesn't come as a surprise)

Also, before r+ing, cc some LLVM folks who might be interested in the implications on LLVM plugins: @rkruppe @nikic (who else?)

Also, cc @tromey. I'm not sure if this will affect LLDB somehow (I suspect it won't).

Thanks a lot for looking into this, @alexcrichton! :heart: