@rkruppe 2018-12-19
Really nice wins! ❤️ 

Regarding impact on LLVM plugins, well, this is also true from the LLVM perspective:

> I feel like the writing has been on the wall for this test for quite some time now.

The new pass manager has never supported dynamically loaded passes and is getting closer and closer to becoming the default. As far as I know nobody stepped up to even design plugins for the new pass manager, let alone see through implementation.