@Centril 2018-12-24
@bors retry

(Included in https://github.com/rust-lang/rust/pull/57094)